pyproject.toml
[project]
name = "mobile-agents"
version = "0.1.0"
description = "Personal agent manager - run AI coding agents asynchronously on a server"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.20.0",
    "pydantic>=2.0.0",
    "greenlet>=3.0.0",
    "pyjwt>=2.0.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.5.0",
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]
